Implement Observer/Viewer and update models
Refactored both server and client code to handle the head/viewer clients
which track the position of the users' head. Players can now look around
in the environment.

Due to the many different considerations of orientation, server now has
a `FrameType` enum to differentiate which kinds of operations to apply
only to the viewer vs the racket.

Still need to correct the acceleration pre-rotations for the viewer
object, and the drift from acceleration is still unsolved.